requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.8
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
mysql-connector-python==9.5.0
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
